(highlight / color overwrite) handle highlight and color overwrite options for all chart types except arrow
