+++ chartTypes/bar/mapping.js
@@ -173,6 +173,7 @@ module.exports = function getMapping() {
       }
     }
   ]
+    .concat(commonMappings.getBarColumnDotHighlightColorOverwriteMappings())
     .concat(commonMappings.getBarDateSeriesHandlingMappings())
     .concat(commonMappings.getBarPrognosisMappings())
     .concat(commonMappings.getBarLabelColorMappings());
