+++ chartTypes/column-stacked/vega-spec.json
@@ -68,7 +68,18 @@
       "name": "colorScale",
       "type": "ordinal",
       "range": {
-        "scheme": "categorical_computed"
+        "scheme": "categorical_computed_normal"
+      },
+      "domain": {
+        "data": "table",
+        "field": "cValue"
+      }
+    },
+    {
+      "name": "colorScaleLight",
+      "type": "ordinal",
+      "range": {
+        "scheme": "categorical_computed_light"
       },
       "domain": {
         "data": "table",
@@ -98,10 +109,34 @@
             "scale": "yScale",
             "field": "endY"
           },
-          "fill": {
-            "scale": "colorScale",
-            "signal": "datum.cValue"
-          }
+          "fill": [
+            {
+              "test": "datum.colorLight && datum.isHighlighted === false",
+              "field": "colorLight"
+            },
+            {
+              "test": "datum.color && datum.isHighlighted === true",
+              "field": "color"
+            },
+            {
+              "test": "datum.isHighlighted === true",
+              "scale": "colorScale",
+              "field": "cValue"
+            },
+            {
+              "test": "datum.isHighlighted === false",
+              "scale": "colorScaleLight",
+              "field": "cValue"
+            },
+            {
+              "test": "datum.color",
+              "field": "color"
+            },
+            {
+              "scale": "colorScale",
+              "field": "cValue"
+            }
+          ]
         }
       }
     }
