+++ chartTypes/commonMappings.js
@@ -282,6 +282,69 @@ function getHeightMappings() {
   ];
 }
 
+function getBarColumnDotHighlightColorOverwriteMappings() {
+  return [
+    {
+      path: "item.options.highlightDataRows",
+      mapToSpec: function(highlightDataRows, spec) {
+        if (highlightDataRows.length === 0) {
+          return;
+        }
+        spec.data[0].values.map(value => {
+          if (highlightDataRows.includes(value.xIndex)) {
+            value.isHighlighted = true;
+          } else {
+            value.isHighlighted = false;
+          }
+        });
+      }
+    },
+    {
+      path: "item.options.highlightDataSeries",
+      mapToSpec: function(highlightDataSeries, spec) {
+        if (highlightDataSeries.length === 0) {
+          return;
+        }
+        spec.data[0].values.map(value => {
+          // if isHighlighted is already set to false, we return here
+          // highlight rows and highlight columns are in an AND relationship to result in a highlighted column
+          if (value.isHighlighted === false) {
+            return;
+          }
+          if (highlightDataSeries.includes(value.cValue)) {
+            value.isHighlighted = true;
+          } else {
+            value.isHighlighted = false;
+          }
+        });
+      }
+    },
+    {
+      path: "item.options.colorOverwritesRows",
+      mapToSpec: function(colorOverwritesRows, spec, mappingData) {
+        // do not handle colorOverwriteRows if we have more than two dataseries.
+        if (mappingData.item.data[0].length > 3) {
+          return;
+        }
+        for (const colorOverwrite of colorOverwritesRows) {
+          //   // if we do not have a valid color or position, ignore this
+          if (!colorOverwrite.color || Number.isNaN(colorOverwrite.position)) {
+            continue;
+          }
+          spec.data[0].values.map(value => {
+            if (value.xIndex === colorOverwrite.position) {
+              value.color = colorOverwrite.color;
+              if (colorOverwrite.colorLight) {
+                value.colorLight = colorOverwrite.colorLight;
+              }
+            }
+          });
+        }
+      }
+    }
+  ];
+}
+
 module.exports = {
   getLineDateSeriesHandlingMappings: getLineDateSeriesHandlingMappings,
   getColumnDateSeriesHandlingMappings: getColumnDateSeriesHandlingMappings,
@@ -290,5 +353,6 @@ module.exports = {
   getBarPrognosisMappings: getBarPrognosisMappings,
   getColumnLabelColorMappings: getColumnLabelColorMappings,
   getBarLabelColorMappings: getBarLabelColorMappings,
-  getHeightMappings: getHeightMappings
+  getHeightMappings: getHeightMappings,
+  getBarColumnDotHighlightColorOverwriteMappings: getBarColumnDotHighlightColorOverwriteMappings
 };
