+++ chartTypes/dotplot/vega-spec.json
@@ -57,7 +57,20 @@
     {
       "name": "colorScale",
       "type": "ordinal",
-      "range": "category",
+      "range": {
+        "scheme": "categorical_computed_normal"
+      },
+      "domain": {
+        "data": "table",
+        "field": "cValue"
+      }
+    },
+    {
+      "name": "colorScaleLight",
+      "type": "ordinal",
+      "range": {
+        "scheme": "categorical_computed_light"
+      },
       "domain": {
         "data": "table",
         "field": "cValue"
@@ -112,9 +125,7 @@
               "transform": [
                 {
                   "type": "aggregate",
-                  "groupby": [
-                    "xValue"
-                  ]
+                  "groupby": ["xValue"]
                 }
               ]
             }
@@ -135,10 +146,6 @@
                     "scale": "xScale",
                     "field": "yValue"
                   },
-                  "fill": {
-                    "scale": "colorScale",
-                    "field": "cValue"
-                  },
                   "stroke": {
                     "value": "white"
                   },
@@ -147,7 +154,35 @@
                   },
                   "size": {
                     "value": 121
-                  }
+                  },
+                  "fill": [
+                    {
+                      "test": "datum.colorLight && datum.isHighlighted === false",
+                      "field": "colorLight"
+                    },
+                    {
+                      "test": "datum.color && datum.isHighlighted === true",
+                      "field": "color"
+                    },
+                    {
+                      "test": "datum.isHighlighted === true",
+                      "scale": "colorScale",
+                      "field": "cValue"
+                    },
+                    {
+                      "test": "datum.isHighlighted === false",
+                      "scale": "colorScaleLight",
+                      "field": "cValue"
+                    },
+                    {
+                      "test": "datum.color",
+                      "field": "color"
+                    },
+                    {
+                      "scale": "colorScale",
+                      "field": "cValue"
+                    }
+                  ]
                 }
               }
             }
