+++ chartTypes/line/vega-spec.json
@@ -30,7 +30,18 @@
       "name": "colorScale",
       "type": "ordinal",
       "range": {
-        "scheme": "categorical_computed"
+        "scheme": "categorical_computed_normal"
+      },
+      "domain": {
+        "data": "table",
+        "field": "cValue"
+      }
+    },
+    {
+      "name": "colorScaleLight",
+      "type": "ordinal",
+      "range": {
+        "scheme": "categorical_computed_light"
       },
       "domain": {
         "data": "table",
